Move COpenGLImage & CVulkanImage to includes
