+++ include/nbl/video/declarations.h
@@ -24,6 +24,8 @@
 #include "nbl/video/CVulkanConnection.h"
 #include "nbl/video/COpenGL_Swapchain.h"
 #include "nbl/video/CVulkanSwapchain.h"
+#include "nbl/video/COpenGLImage.h"
+#include "nbl/video/CVulkanImage.h"
 #include "nbl/video/surface/CSurfaceGL.h"
 #include "nbl/video/surface/CSurfaceVulkan.h"
